requirement.txt
networkx==3.3
numpy==1.26.4
Pillow==9.0.1
matplotlib==3.9.1
scikit-learn==1.5.1